+++ frontend/src/modules/admin.transactions/TransactionService.js
@@ -12,6 +12,21 @@ export default class TransactionService {
         return this.Restangular.one('transaction').customPOST({transaction: newTransaction})
     }
 
+    postImportTransaction(file) {
+        var formData = new FormData();
+        formData.append('file[file]', file);
+
+        return this.Restangular
+            .one('admin')
+            .one('transaction')
+            .one('import')
+            .withHttpConfig({
+                transformRequest: angular.identity,
+                timeout: 0
+            })
+            .customPOST(formData, '', undefined, {'Content-Type': undefined})
+    }
+
     postAssign(linked) {
         return this.Restangular
             .one('admin')