+++ frontend/src/modules/admin.transfers/TransferController.js
@@ -81,7 +81,8 @@ export default class TransferController {
             transferList: true,
             coverLoader: true,
             addTransfer: false,
-            cancelTransfer: false
+            cancelTransfer: false,
+            importTransfer: false
         }
     }
 
@@ -123,6 +124,19 @@ export default class TransferController {
 
         self.$scope.showSpendPoints = false;
     }
+
+    openImportModal() {
+        let self = this;
+
+        self.$scope.showTransferImportModal = true;
+    }
+
+    closeImportModal() {
+        let self = this;
+
+        self.$scope.showTransferImportModal = false;
+    }
+
 
     getData() {
         let self = this;
@@ -267,6 +281,42 @@ export default class TransferController {
                     self.loaderStates.cancelTransfer = false;
                 }
             )
+    }
+
+    importTransfer(file) {
+        let self = this;
+
+        if (file) {
+            self.loaderStates.importTransfer = true;
+            self.TransferService.postImportTransfer(file)
+                .then(
+                    res => {
+                        if (res.totalProcessed == 0) {
+                            let message = self.$filter('translate')('xhr.import.no_data');
+                            self.Flash.create('warning', message);
+                        }
+                        else if (res.totalFailed == 0) {
+                            let message = self.$filter('translate')('xhr.import.success', {processed: res.totalProcessed});
+                            self.Flash.create('success', message);
+                        } else {
+                            let message = self.$filter('translate')('xhr.import.warning',
+                                {processed: res.totalProcessed, success: res.totalSuccess, failed: res.totalFailed}
+                            );
+                            self.Flash.create('warning', message);
+                        }
+
+                        self.$scope.importTransactionModal = false;
+                        self.loaderStates.importTransfer = false;
+                        self.closeImportModal();
+                        self.tableParams.reload();
+                    },
+                    res => {
+                        let message = self.$filter('translate')('xhr.import.error');
+                        self.Flash.create('danger', message);
+                        self.loaderStates.importTransfer = false;
+                    }
+                );
+        }
     }
 }