+++ frontend/src/modules/client.earning-rules/CustomerEarningRulesController.js
@@ -1,5 +1,5 @@
 export default class CustomerEarningRulesController {
-    constructor($scope, $state, AuthService, CustomerEarningRulesService, Flash,ParamsMap, $stateParams, EditableMap) {
+    constructor($scope, $state, AuthService, CustomerEarningRulesService, Flash,ParamsMap, $stateParams, EditableMap, DataService) {
         if (!AuthService.isGranted('ROLE_PARTICIPANT')) {
             $state.go('customer-login')
         }
@@ -10,6 +10,7 @@ export default class CustomerEarningRulesController {
         this.ParamsMap = ParamsMap;
         this.EditableMap = EditableMap;
         this.$scope.loader = true;
+        this.config = DataService.getConfig();
     }
 
     getRules() {
@@ -20,9 +21,27 @@ export default class CustomerEarningRulesController {
                 self.$scope.loader = false;
                 self.$scope.rules = res.earningRules;
                 self.$scope.currency = res.currency;
+                angular.forEach(self.$scope.rules, function(rule, key) {
+                    self.CustomerEarningRulesService.getEarningRuleImage(rule.earningRuleId)
+                        .then(
+                            res => {
+                                rule.earningRuleImagePath = true;
+                            }
+                        )
+                        .catch(
+                            err => {
+
+                                rule.earningRuleImagePath = false;
+                            }
+                        );
+                });
             }
         )
     }
+
+    generatePhotoRoute(earningRuleId) {
+        return this.config.apiUrl + '/earningRule/' + earningRuleId + '/photo'
+    }
 }
 
-CustomerEarningRulesController.$inject = ['$scope', '$state', 'AuthService', 'CustomerEarningRulesService', 'Flash', 'ParamsMap', '$stateParams', 'EditableMap'];
+CustomerEarningRulesController.$inject = ['$scope', '$state', 'AuthService', 'CustomerEarningRulesService', 'Flash', 'ParamsMap', '$stateParams', 'EditableMap', 'DataService'];