+++ frontend/src/modules/client.earning-rules/CustomerEarningRulesService.js
@@ -7,6 +7,17 @@ export default class CustomerEarningRulesService {
     getRules($id) {
         return this.Restangular.one('customer').one('earningRule').get();
     }
+
+    /**
+     * Calls for earning rule image
+     *
+     * @method getEarningRuleImage
+     * @param {Integer} earningRuleId
+     * @returns {Promise}
+     */
+    getEarningRuleImage(earningRuleId) {
+        return this.Restangular.one('earningRule', earningRuleId).one('photo').get();
+    }
 }
 
 CustomerEarningRulesService.$inject = ['Restangular', 'EditableMap'];